+++ src/tiposProductos/Comida.java
@@ -1,7 +1,7 @@
 package tiposProductos;
 public class Comida extends Productos {
 
-    public Comida(int id, String nombre, double precio, int stock, boolean alergeno) {
+    public Comida(int id, String nombre, float precio, int stock, boolean alergeno) {
         super(id, nombre, precio, stock, alergeno);
         // Constructor que me ha obligado a poner por tener herencia
     }
